+++ crates/opsml_server/src/core/cards/utils.rs
@@ -118,6 +118,7 @@ pub async fn insert_card_into_db(
                 client_card.service_type,
                 client_card.metadata,
                 client_card.deployment,
+                client_card.service_config,
                 client_card.username,
             );
             ServerCard::Service(server_card)
