+++ crates/opsml_service/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod error;
 pub mod service;
 
-pub use service::{Card, DriftConfig, ServiceConfig, ServiceSpec};
+pub use service::ServiceSpec;
