+++ crates/opsml_sql/src/postgres/sql/card/mod.rs
@@ -360,6 +360,7 @@ impl CardLogicTrait for CardLogicPostgresClient {
                         .bind(&record.service_type)
                         .bind(&record.metadata)
                         .bind(&record.deployment)
+                        .bind(&record.service_config)
                         .execute(&self.pool)
                         .await?;
                     Ok(())
@@ -544,6 +545,7 @@ impl CardLogicTrait for CardLogicPostgresClient {
                         .bind(&record.service_type)
                         .bind(&record.metadata)
                         .bind(&record.deployment)
+                        .bind(&record.service_config)
                         .bind(&record.uid)
                         .execute(&self.pool)
                         .await?;
