+++ crates/opsml_sql/src/sqlite/sql/card/insert_servicecard.sql
@@ -14,7 +14,8 @@ INSERT INTO opsml_service_registry (
     opsml_version, 
     service_type,
     metadata,
-    deployment
+    deployment,
+    service_config
     ) 
     VALUES 
-    (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);
+    (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);
