+++ crates/opsml_sql/src/sqlite/sql/card/mod.rs
@@ -362,6 +362,7 @@ impl CardLogicTrait for CardLogicSqliteClient {
                         .bind(&record.service_type)
                         .bind(&record.metadata)
                         .bind(&record.deployment)
+                        .bind(&record.service_config)
                         .execute(&self.pool)
                         .await?;
                     Ok(())
@@ -546,6 +547,7 @@ impl CardLogicTrait for CardLogicSqliteClient {
                         .bind(&record.service_type)
                         .bind(&record.metadata)
                         .bind(&record.deployment)
+                        .bind(&record.service_config)
                         .bind(&record.uid)
                         .execute(&self.pool)
                         .await?;
