+++ crates/opsml_sql/src/sqlite/sql/card/update_servicecard.sql
@@ -11,5 +11,6 @@ UPDATE opsml_service_registry SET
     opsml_version = ?,
     service_type = ?,
     metadata = ?,
-    deployment = ?
+    deployment = ?,
+    service_config = ?
 WHERE uid = ?;
