+++ crates/opsml_types/src/contracts/mod.rs
@@ -6,6 +6,7 @@ pub mod experiment;
 pub mod file;
 pub mod mcp;
 pub mod scouter;
+pub mod service;
 pub mod traits;
 
 pub use artifact::*;
@@ -15,4 +16,5 @@ pub use experiment::*;
 pub use file::*;
 pub use mcp::*;
 pub use scouter::*;
+pub use service::*;
 pub use traits::*;
